+++ PhzQtUI/src/lib/DialogFilterShiftGridGeneration.cpp
@@ -77,9 +77,10 @@ DialogFilterShiftGridGeneration::DialogFilterShiftGridGeneration(QWidget* parent
 DialogFilterShiftGridGeneration::~DialogFilterShiftGridGeneration() {}
 
 void DialogFilterShiftGridGeneration::setValues(
-    std::string grid_name, const std::map<std::string, boost::program_options::variable_value>& config) {
+    std::string grid_name, const std::map<std::string, boost::program_options::variable_value>& config, double pp_norm) {
   ui->label_name->setText(QString::fromStdString(grid_name));
   m_config = config;
+  m_pp_norm = pp_norm;
 }
 
 std::string DialogFilterShiftGridGeneration::runFunction() {
@@ -103,11 +104,15 @@ std::string DialogFilterShiftGridGeneration::runFunction() {
     auto shift_sampling = config_manager.template getConfiguration<FilterVariationConfig>().getSampling();
     auto cosmology = config_manager.template getConfiguration<CosmologicalParameterConfig>().getCosmologicalParam();
     auto lum_filter_name = config_manager.getConfiguration<ModelNormalizationConfig>().getNormalizationFilter();
+    auto lum_pp_filter_name = config_manager.getConfiguration<ModelNormalizationConfig>().getPpNormalizationFilter();
     auto sun_sed_name    = config_manager.getConfiguration<ModelNormalizationConfig>().getReferenceSolarSed();
 
     auto normalizer_functor =
         Euclid::PhzModeling::NormalizationFunctorFactory::NormalizationFunctorFactory::GetFunction(
             filter_provider, lum_filter_name, sed_provider, sun_sed_name);
+    auto normalizer_pp_functor =
+        Euclid::PhzModeling::NormalizationFunctorFactory::NormalizationFunctorFactory::GetFunction(
+            filter_provider, lum_pp_filter_name, sed_provider, sun_sed_name);
 
     std::map<std::string, PhzDataModel::PhotometryGrid> result_map{};
 
@@ -126,8 +131,14 @@ std::string DialogFilterShiftGridGeneration::runFunction() {
       }
     };
 
-    PhzFilterVariation::FilterVariationSingleGridCreator grid_creator{sed_provider, reddening_provider, filter_provider,
-                                                                      igm_abs_func, normalizer_functor, shift_sampling};
+    PhzFilterVariation::FilterVariationSingleGridCreator grid_creator{sed_provider, 
+                                                                      reddening_provider, 
+                                                                      filter_provider,
+                                                                      igm_abs_func, 
+                                                                      normalizer_functor,
+																	  normalizer_pp_functor, 
+																	  m_pp_norm,
+																	  shift_sampling};
     size_t                                               already_done = 0;
     for (auto& grid_pair : model_phot_grid.region_axes_map) {
       SparseProgressReporter reporter{monitor_function, already_done, total};